pyproject.toml
[project]
name = "scr"
version = "0.1.0"
description = ""
authors = [
    {name = "MR-RITZY",email = "mrboffin01@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi (>=0.119.0,<0.120.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "numerizer (>=0.2.4,<0.3.0)",
    "lark (>=1.3.0,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
